Resize using inverseFFT and default shrinker/expander.
This interpolates in the spatial domain, which is not the best.
The other approach to resize in the spatial domain:
If you want to down/up sample by 2.
Shrink: remove odd (or even) samples.
Expand: add zero voxel between samples.
This is done in OTB (not working with 3D though)